refactor: Layout 관련 스타일 분리
