+++ src/components/InningPlot/InningPlot.styles.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { FlexColumnContainer } from '@components/Layout/Layout';
+import { FlexColumnContainer } from '@components/Layout/Layout.styles';
 
 export const InningPlotContainer = styled(FlexColumnContainer)({
     justifyContent: 'flex-start',
