+++ src/components/MobileHeader/MobileHeader.styles.tsx
@@ -2,7 +2,7 @@ import { NavLink } from 'react-router-dom';
 
 import styled from '@emotion/styled';
 
-import { FlexRowContainer } from '@components/Layout/Layout';
+import { FlexRowContainer } from '@components/Layout/Layout.styles';
 
 export const MobileHeaderContainer = styled(FlexRowContainer)({
     justifyContent: 'flex-start',
