+++ src/components/SingleCalendarSection/SingleCalendarSection.styles.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { FlexColumnContainer } from '@components/Layout/Layout';
+import { FlexColumnContainer } from '@components/Layout/Layout.styles';
 
 export const SingleCalendarSectionContainer = styled(FlexColumnContainer)({
     alignItems: 'flex-start',
