+++ src/components/StrikeZone/StrikeZone.styles.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { FlexRowContainer } from '@components/Layout/Layout';
+import { FlexRowContainer } from '@components/Layout/Layout.styles';
 
 export const StrikeZoneRow = styled(FlexRowContainer)({
     justifyContent: 'flex-start',
