+++ src/pages/GameDetailPage/components/GameDataSection/GameDataSection.styles.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { FlexColumnContainer, FlexRowContainer } from '@components/Layout/Layout';
+import { FlexColumnContainer, FlexRowContainer } from '@components/Layout/Layout.styles';
 
 export const GameDataSectionContainer = styled(FlexColumnContainer)({
     justifyContent: 'flex-start',
