+++ src/pages/PlayerDetailPage/components/PlayerGameList/PlayerGameList.styles.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { FlexColumnContainer } from '@components/Layout/Layout';
+import { FlexColumnContainer } from '@components/Layout/Layout.styles';
 
 export const PlayerGameListContainer = styled(FlexColumnContainer)({
     gap: '0.5rem',
