+++ src/pages/PlayerDetailPage/components/PlayerStrikeZone/PlayerStrikeZone.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import { SectionTitle } from '@components/Layout/Layout';
+import { SectionTitle } from '@components/Layout/Layout.styles';
 import { Loading } from '@components/Loading/Loading';
 
 import { useGetAtbatsByPlayerSlug } from '@hooks/@query/player/useGetAtbatsByPlayerSlug';
