+++ src/pages/PlayerListPage/PlayerListPage.styles.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { ColumnPageContainer } from '@components/Layout/Layout';
+import { ColumnPageContainer } from '@components/Layout/Layout.styles';
 
 export const PlayerListContainer = styled(ColumnPageContainer)({
     justifyContent: 'flex-start',
