+++ src/pages/TeamDetailPage/components/TeamGameList/TeamGameList.styles.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { FlexColumnContainer } from '@components/Layout/Layout';
+import { FlexColumnContainer } from '@components/Layout/Layout.styles';
 
 import { breakpoints } from '@styles/theme';
