+++ src/pages/TeamDetailPage/components/TeamRoster/TeamRoster.styles.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { FlexColumnContainer } from '@components/Layout/Layout';
+import { FlexColumnContainer } from '@components/Layout/Layout.styles';
 
 export const TeamRosterContainer = styled(FlexColumnContainer)({
     gap: '1rem',
