+++ src/pages/UmpireDetailPage/UmpireDetailPage.styles.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { ColumnPageContainer, FlexColumnContainer } from '@components/Layout/Layout';
+import { ColumnPageContainer, FlexColumnContainer } from '@components/Layout/Layout.styles';
 
 import { breakpoints } from '@styles/theme';
