+++ src/pages/UmpireListPage/UmpireListPage.styles.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { ColumnPageContainer } from '@components/Layout/Layout';
+import { ColumnPageContainer } from '@components/Layout/Layout.styles';
 
 export const UmpireListPageContainer = styled(ColumnPageContainer)({
     gap: '1rem',
