+++ src/pages/UmpireListPage/components/UmpireListItem/UmpireListItem.styles.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import { FlexColumnContainer } from '@components/Layout/Layout';
+import { FlexColumnContainer } from '@components/Layout/Layout.styles';
 
 import { breakpoints } from '@styles/theme';
